ADD: create route file for product routes
